requirements.txt
pythonocc-core>=7.7.0
gmsh>=4.11.0
trimesh[easy]>=3.23.0
numpy>=1.24.0
scipy>=1.10.0
networkx>=3.0
